requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=hd4dd3e8_0
altair=4.1.0=py_1
appdirs=1.4.4=pyhd3eb1b0_0
arrow-cpp=11.0.0=py39h613000e_0
asttokens=2.0.5=pyhd3eb1b0_0
attrs=22.1.0=py39h06a4308_0
autopep8=2.0.2=pyhd8ed1ab_0
aws-c-common=0.4.57=he6710b0_1
aws-c-event-stream=0.1.6=h2531618_5
aws-checksums=0.1.9=he6710b0_0
aws-sdk-cpp=1.8.185=hce553d0_0
backcall=0.2.0=pyhd3eb1b0_0
beautifulsoup4=4.12.2=py39h06a4308_0
blas=1.0=mkl
bleach=6.0.0=pypi_0
blinker=1.4=py39h06a4308_0
boltons=23.0.0=py39h06a4308_0
boost-cpp=1.73.0=h7f8727e_12
bottleneck=1.3.5=py39h7deecbd_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py39h27cfd23_1003
build=0.10.0=pypi_0
bzip2=1.0.8=h7b6447c_0
c-ares=1.19.0=h5eee18b_0
ca-certificates=2023.01.10=h06a4308_0
cachetools=4.2.2=pyhd3eb1b0_0
certifi=2022.12.7=py39h06a4308_0
cffi=1.15.1=py39h5eee18b_3
chardet=4.0.0=py39h06a4308_1003
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py39h06a4308_0
comm=0.1.2=py39h06a4308_0
commonmark=0.9.1=pyhd3eb1b0_0
conda=23.3.1=py39h06a4308_0
conda-build=3.24.0=py39h06a4308_0
conda-package-handling=2.0.2=py39h06a4308_0
conda-package-streaming=0.7.0=py39h06a4308_0
contourpy=1.0.5=py39hdb19cb5_0
convertdate=2.3.2=pyhd3eb1b0_0
cryptography=39.0.1=py39h9ce1e76_0
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.33=py39h6a678d5_0
d2c-mp-sales-forecaster=0.0.2=pypi_0
dataclasses=0.8=pyh6d0b6a4_7
debugpy=1.5.1=py39h295c915_0
decorator=5.1.1=pyhd3eb1b0_0
docutils=0.19=pypi_0
entrypoints=0.4=py39h06a4308_0
ephem=4.1.2=py39h7f8727e_0
executing=0.8.3=pyhd3eb1b0_0
filelock=3.9.0=py39h06a4308_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
future=0.18.3=py39h06a4308_0
gflags=2.2.2=he6710b0_0
giflib=5.2.1=h5eee18b_3
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.30=py39h06a4308_0
glob2=0.7=pyhd3eb1b0_0
glog=0.5.0=h2531618_0
grpc-cpp=1.46.1=h33aed49_1
hierarchicalforecast=0.3.0=pyhd8ed1ab_0
hijri-converter=2.2.2=pyhd3eb1b0_0
holidays=0.18=py39h06a4308_0
icu=58.2=he6710b0_3
idna=3.4=py39h06a4308_0
importlib-metadata=6.0.0=py39h06a4308_0
importlib_metadata=6.0.0=hd3eb1b0_0
importlib_resources=5.2.0=pyhd3eb1b0_1
iniconfig=1.1.1=pyhd3eb1b0_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.19.2=py39hb070fc8_0
ipython=8.12.0=py39h06a4308_0
ipywidgets=8.0.4=py39h06a4308_0
jaraco-classes=3.2.3=pypi_0
jedi=0.18.1=py39h06a4308_1
jeepney=0.8.0=pypi_0
jinja2=3.1.2=py39h06a4308_0
joblib=1.1.1=py39h06a4308_0
jpeg=9e=h5eee18b_1
jsonpatch=1.32=pyhd3eb1b0_0
jsonpointer=2.1=pyhd3eb1b0_0
jsonschema=4.17.3=py39h06a4308_0
jupyter_client=8.1.0=py39h06a4308_0
jupyter_core=5.3.0=py39h06a4308_0
jupyterlab_widgets=3.0.5=py39h06a4308_0
keyring=23.13.1=pypi_0
kiwisolver=1.4.4=py39h6a678d5_0
korean_lunar_calendar=0.2.1=pyhd3eb1b0_0
krb5=1.19.4=h568e23c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libarchive=3.6.2=hb4bd9a0_1
libboost=1.73.0=h28710b8_12
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcurl=7.88.1=h91b91d3_0
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libev=4.33=h7f8727e_1
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
liblief=0.12.3=h6a678d5_0
libllvm11=11.1.0=h9e868ea_6
libnghttp2=1.46.0=hce63b2e_0
libpng=1.6.39=h5eee18b_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.10.0=h8f2d780_0
libstdcxx-ng=11.2.0=h1234567_1
libthrift=0.15.0=hcc01f38_0
libtiff=4.5.0=h6a678d5_2
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxml2=2.10.3=hcbfbd50_0
llvmlite=0.39.1=py39he621ea3_0
lunarcalendar=0.0.9=pyhd3eb1b0_0
lz4-c=1.9.4=h6a678d5_0
markupsafe=2.1.1=py39h7f8727e_0
matplotlib-base=3.7.1=py39h417a72b_1
matplotlib-inline=0.1.6=py39h06a4308_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py39h7f8727e_0
mkl_fft=1.3.1=py39hd3c417c_0
mkl_random=1.2.2=py39h51133e4_0
more-itertools=9.1.0=pypi_0
munkres=1.1.4=py_0
ncurses=6.4=h6a678d5_0
nest-asyncio=1.5.6=py39h06a4308_0
neuralprophet=0.4.2=py39h06a4308_0
ninja=1.10.2=h06a4308_5
ninja-base=1.10.2=hd09550d_5
numba=0.56.4=py39h417a72b_0
numexpr=2.8.4=py39he184ba9_0
numpy=1.23.5=py39h14f4228_0
numpy-base=1.23.5=py39h31eccc5_0
openssl=1.1.1t=h7f8727e_0
orc=1.7.4=hb3bc3d3_1
packaging=23.0=py39h06a4308_0
pandas=1.5.3=py39h417a72b_0
pandoc=2.12=h06a4308_3
parso=0.8.3=pyhd3eb1b0_0
patch=2.7.6=h7b6447c_1001
patchelf=0.17.2=h6a678d5_0
patsy=0.5.3=py39h06a4308_0
pep517=0.12.0=py39h06a4308_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py39h6a678d5_0
pip=23.0.1=py39h06a4308_0
pkginfo=1.9.6=py39h06a4308_0
platformdirs=2.5.2=py39h06a4308_0
plotly=5.9.0=py39h06a4308_0
pluggy=1.0.0=py39h06a4308_1
pooch=1.4.0=pyhd3eb1b0_0
prompt-toolkit=3.0.36=py39h06a4308_0
protobuf=3.20.3=py39h6a678d5_0
psutil=5.9.0=py39h5eee18b_0
psycopg2=2.9.3=py39h37d81fd_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
py=1.11.0=pyhd3eb1b0_0
py-lief=0.12.3=py39h6a678d5_0
pyarrow=11.0.0=py39h992f0b0_0
pycodestyle=2.10.0=py39h06a4308_0
pycosat=0.6.4=py39h5eee18b_0
pycparser=2.21=pyhd3eb1b0_0
pydeck=0.7.1=py39h06a4308_0
pygments=2.11.2=pyhd3eb1b0_0
pymeeus=0.5.11=pyhd3eb1b0_1
pympler=0.9=py_0
pyopenssl=23.0.0=py39h06a4308_0
pypandoc=1.6.3=pyhd3eb1b0_1
pyparsing=3.0.9=py39h06a4308_0
pyproject-hooks=1.0.0=pypi_0
pyrsistent=0.18.0=py39heee7806_0
pysocks=1.7.1=py39h06a4308_0
pytest=7.1.2=py39h06a4308_0
pytest-mock=3.10.0=pyhd8ed1ab_0
python=3.9.16=h7a1cb2a_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python-libarchive-c=2.9=pyhd3eb1b0_1
python_abi=3.9=2_cp39
pytorch=1.12.1=cpu_py39hb1f1ab4_1
pytz=2022.7=py39h06a4308_0
pyyaml=6.0=py39h5eee18b_1
pyzmq=23.2.0=py39h6a678d5_0
quadprog=0.1.11=py39hf939315_1
re2=2022.04.01=h295c915_0
readline=8.2=h5eee18b_0
readme-renderer=37.3=pypi_0
requests=2.28.1=py39h06a4308_1
requests-toolbelt=0.10.1=pypi_0
rfc3986=2.0.0=pypi_0
rich=12.5.1=py39h06a4308_0
ripgrep=13.0.0=hbdeaff8_0
ruamel.yaml=0.17.21=py39h5eee18b_0
ruamel.yaml.clib=0.2.6=py39h5eee18b_1
scikit-learn=1.2.2=py39h6a678d5_0
scipy=1.10.1=py39h14f4228_0
secretstorage=3.3.3=pypi_0
semver=2.13.0=pyhd3eb1b0_0
setuptools=66.0.0=py39h06a4308_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
snappy=1.1.9=h295c915_0
soupsieve=2.4=py39h06a4308_0
sqlite=3.41.2=h5eee18b_0
stack_data=0.2.0=pyhd3eb1b0_0
statsmodels=0.13.5=py39h7deecbd_1
streamlit=1.19.0=pyhd8ed1ab_0
tbb=2021.8.0=hdb19cb5_0
tenacity=8.0.1=py39h06a4308_1
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
tomli=2.0.1=py39h06a4308_0
toolz=0.12.0=py39h06a4308_0
torch-lr-finder=0.2.1=pyhd3eb1b0_0
tornado=6.2=py39h5eee18b_0
tqdm=4.65.0=py39hb070fc8_0
traitlets=5.7.1=py39h06a4308_0
twine=4.0.2=pypi_0
typing-extensions=4.5.0=py39h06a4308_0
typing_extensions=4.5.0=py39h06a4308_0
tzdata=2023c=h04d1e81_0
tzlocal=2.1=py39h06a4308_1
urllib3=1.26.15=py39h06a4308_0
utf8proc=2.6.1=h27cfd23_0
validators=0.18.2=pyhd3eb1b0_0
watchdog=2.1.6=py39h06a4308_0
wcwidth=0.2.5=pyhd3eb1b0_0
webencodings=0.5.1=pypi_0
wheel=0.38.4=py39h06a4308_0
widgetsnbextension=4.0.5=py39h06a4308_0
xz=5.2.10=h5eee18b_1
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zipp=3.11.0=py39h06a4308_0
zlib=1.2.13=h5eee18b_0
zstandard=0.19.0=py39h5eee18b_0
zstd=1.5.5=hc292b87_0
